+++ tools/globals.js
@@ -82,6 +82,7 @@ Minimap Minimap_ZoomIn GetMinimapShape
 ITEM_QUALITY_POOR ITEM_QUALITY_COMMON ITEM_QUALITY_UNCOMMON ITEM_QUALITY_RARE
 ITEM_QUALITY_EPIC OKAY YES NO CANCEL ACCEPT CLOSE
 ITEM_UNIQUE ITEM_UNIQUE_MULTIPLE ITEM_UNIQUE_EQUIPPABLE ITEM_SPELL_KNOWN
+EMPTY_SOCKET_RED EMPTY_SOCKET_YELLOW EMPTY_SOCKET_BLUE EMPTY_SOCKET_META EMPTY_SOCKET_PRISMATIC
 GetNumSkillLines GetSkillLineInfo
 strtrim strsplit strjoin strlower strupper strsub strlen strfind strmatch strrep
 tinsert tremove twipe wipe format gsub gmatch max min abs floor ceil